updating fork

+++ src/services/cognitoService.ts
@@ -1,18 +1,18 @@
+import type { Dirent } from "node:fs";
+import fs from "node:fs/promises";
+import * as path from "node:path";
 import mergeWith from "lodash.mergewith";
-import * as path from "path";
 import { ResourceNotFoundError } from "../errors";
-import { UserPoolDefaults } from "../server/config";
-import { AppClient } from "./appClient";
-import { Clock } from "./clock";
-import { Context } from "./context";
-import { DataStore } from "./dataStore/dataStore";
-import { DataStoreFactory } from "./dataStore/factory";
-import {
+import type { UserPoolDefaults } from "../server/config";
+import type { AppClient } from "./appClient";
+import type { Context } from "./context";
+import type { DataStore } from "./dataStore/dataStore";
+import type { DataStoreFactory } from "./dataStore/factory";
+import type {
   UserPool,
   UserPoolService,
   UserPoolServiceFactory,
 } from "./userPoolService";
-import fs from "fs/promises";
 
 const CLIENTS_DATABASE_NAME = "clients";
 
@@ -267,11 +267,11 @@ export interface CognitoService {
   getUserPool(ctx: Context, userPoolId: string): Promise<UserPoolService>;
   getUserPoolForClientId(
     ctx: Context,
-    clientId: string
+    clientId: string,
   ): Promise<UserPoolService>;
   listAppClients(
     ctx: Context,
-    userPoolId: string
+    userPoolId: string,
   ): Promise<readonly AppClient[]>;
   listUserPools(ctx: Context): Promise<readonly UserPool[]>;
 }
@@ -279,26 +279,30 @@ export interface CognitoService {
 export interface CognitoServiceFactory {
   create(
     ctx: Context,
-    userPoolDefaultConfig: UserPoolDefaults
+    userPoolDefaultConfig: UserPoolDefaults,
   ): Promise<CognitoService>;
+}
+
+class NotInitializedError extends Error {
+  public constructor() {
+    super("Not initialized, CognitoServiceImpl.init() must be called first");
+  }
 }
 
 export class CognitoServiceImpl implements CognitoService {
   private readonly clients: DataStore;
-  private readonly clock: Clock;
   private readonly userPoolServiceFactory: UserPoolServiceFactory;
   private readonly dataDirectory: string;
   private readonly userPoolDefaultConfig: UserPoolDefaults;
+  private userPools: UserPoolService[] | undefined;
 
   public constructor(
     dataDirectory: string,
     clients: DataStore,
-    clock: Clock,
     userPoolDefaultConfig: UserPoolDefaults,
-    userPoolServiceFactory: UserPoolServiceFactory
+    userPoolServiceFactory: UserPoolServiceFactory,
   ) {
     this.clients = clients;
-    this.clock = clock;
     this.dataDirectory = dataDirectory;
     this.userPoolDefaultConfig = userPoolDefaultConfig;
     this.userPoolServiceFactory = userPoolServiceFactory;
@@ -306,9 +310,14 @@ export class CognitoServiceImpl implements CognitoService {
 
   public async createUserPool(
     ctx: Context,
-    userPool: UserPool
+    userPool: UserPool,
   ): Promise<UserPool> {
     ctx.logger.debug("CognitoServiceImpl.createUserPool");
+
+    if (!this.userPools) {
+      throw new NotInitializedError();
+    }
+
     const service = await this.userPoolServiceFactory.create(
       ctx,
       this.clients,
@@ -316,9 +325,11 @@ export class CognitoServiceImpl implements CognitoService {
         {},
         USER_POOL_AWS_DEFAULTS,
         this.userPoolDefaultConfig,
-        userPool
-      )
+        userPool,
+      ),
     );
+
+    this.userPools.push(service);
 
     return service.options;
   }
@@ -326,43 +337,63 @@ export class CognitoServiceImpl implements CognitoService {
   public async deleteUserPool(ctx: Context, userPool: UserPool): Promise<void> {
     ctx.logger.debug(
       { userPoolId: userPool.Id },
-      "CognitoServiceImpl.deleteUserPool"
+      "CognitoServiceImpl.deleteUserPool",
     );
+
+    if (!this.userPools) {
+      throw new NotInitializedError();
+    }
+
     await fs.rm(path.join(this.dataDirectory, `${userPool.Id}.json`));
+    this.userPools = this.userPools.filter((x) => x.options.Id !== userPool.Id);
   }
 
-  public async getUserPool(
+  public getUserPool(
     ctx: Context,
-    userPoolId: string
+    userPoolId: string,
   ): Promise<UserPoolService> {
     ctx.logger.debug({ userPoolId }, "CognitoServiceImpl.getUserPool");
-    return this.userPoolServiceFactory.create(ctx, this.clients, {
-      ...USER_POOL_AWS_DEFAULTS,
-      ...this.userPoolDefaultConfig,
-      Id: userPoolId,
-    });
+    if (!this.userPools) {
+      throw new NotInitializedError();
+    }
+
+    const userPool = this.userPools.find((x) => x.options.Id === userPoolId);
+    if (!userPool) {
+      throw new ResourceNotFoundError(`User Pool ${userPoolId} not found`);
+    }
+
+    return Promise.resolve(userPool);
   }
 
   public async getUserPoolForClientId(
     ctx: Context,
-    clientId: string
+    clientId: string,
   ): Promise<UserPoolService> {
     ctx.logger.debug({ clientId }, "CognitoServiceImpl.getUserPoolForClientId");
-    const appClient = await this.getAppClient(ctx, clientId);
-    if (!appClient) {
-      throw new ResourceNotFoundError();
+    if (!this.userPools) {
+      throw new NotInitializedError();
     }
 
-    return this.userPoolServiceFactory.create(ctx, this.clients, {
-      ...USER_POOL_AWS_DEFAULTS,
-      ...this.userPoolDefaultConfig,
-      Id: appClient.UserPoolId,
-    });
+    const appClient = await this.getAppClient(ctx, clientId);
+    if (!appClient) {
+      throw new ResourceNotFoundError(`App Client ${clientId} not found`);
+    }
+
+    const userPool = this.userPools.find(
+      (x) => x.options.Id === appClient.UserPoolId,
+    );
+    if (!userPool) {
+      throw new ResourceNotFoundError(
+        `User Pool ${appClient.UserPoolId} not found`,
+      );
+    }
+
+    return userPool;
   }
 
   public async getAppClient(
     ctx: Context,
-    clientId: string
+    clientId: string,
   ): Promise<AppClient | null> {
     ctx.logger.debug({ clientId }, "CognitoServiceImpl.getAppClient");
     return this.clients.get(ctx, ["Clients", clientId]);
@@ -370,79 +401,90 @@ export class CognitoServiceImpl implements CognitoService {
 
   public async listAppClients(
     ctx: Context,
-    userPoolId: string
+    userPoolId: string,
   ): Promise<readonly AppClient[]> {
     ctx.logger.debug({ userPoolId }, "CognitoServiceImpl.listAppClients");
     const clients = await this.clients.get<Record<string, AppClient>>(
       ctx,
       "Clients",
-      {}
+      {},
     );
 
     return Object.values(clients).filter((x) => x.UserPoolId === userPoolId);
   }
 
-  public async listUserPools(ctx: Context): Promise<readonly UserPool[]> {
+  public listUserPools(ctx: Context): Promise<readonly UserPool[]> {
     ctx.logger.debug("CognitoServiceImpl.listUserPools");
+    if (!this.userPools) {
+      throw new NotInitializedError();
+    }
+
+    return Promise.resolve(this.userPools.map((x) => x.options));
+  }
+
+  public async init(ctx: Context) {
+    function userPoolIdFromDirent(x: Dirent) {
+      return path.basename(x.name, path.extname(x.name));
+    }
+
+    ctx.logger.debug("CognitoServiceImpl.init");
     const entries = await fs.readdir(this.dataDirectory, {
       withFileTypes: true,
     });
 
-    return Promise.all(
+    this.userPools = await Promise.all(
       entries
         .filter(
           (x) =>
             x.isFile() &&
             path.extname(x.name) === ".json" &&
-            path.basename(x.name, path.extname(x.name)) !==
-              CLIENTS_DATABASE_NAME
+            userPoolIdFromDirent(x) !== CLIENTS_DATABASE_NAME,
         )
-        .map(async (x) => {
-          const userPool = await this.getUserPool(
-            ctx,
-            path.basename(x.name, path.extname(x.name))
-          );
-
-          return userPool.options;
-        })
+        .map(async (x) =>
+          this.userPoolServiceFactory.create(ctx, this.clients, {
+            ...USER_POOL_AWS_DEFAULTS,
+            ...this.userPoolDefaultConfig,
+            Id: userPoolIdFromDirent(x),
+          }),
+        ),
     );
   }
 }
 
 export class CognitoServiceFactoryImpl implements CognitoServiceFactory {
   private readonly dataDirectory: string;
-  private readonly clock: Clock;
   private readonly dataStoreFactory: DataStoreFactory;
   private readonly userPoolServiceFactory: UserPoolServiceFactory;
 
   public constructor(
     dataDirectory: string,
-    clock: Clock,
     dataStoreFactory: DataStoreFactory,
-    userPoolServiceFactory: UserPoolServiceFactory
+    userPoolServiceFactory: UserPoolServiceFactory,
   ) {
     this.dataDirectory = dataDirectory;
-    this.clock = clock;
     this.dataStoreFactory = dataStoreFactory;
     this.userPoolServiceFactory = userPoolServiceFactory;
   }
 
   public async create(
     ctx: Context,
-    userPoolDefaultConfig: UserPoolDefaults
+    userPoolDefaultConfig: UserPoolDefaults,
   ): Promise<CognitoService> {
     const clients = await this.dataStoreFactory.create(
       ctx,
       CLIENTS_DATABASE_NAME,
-      { Clients: {} }
+      { Clients: {} },
     );
 
-    return new CognitoServiceImpl(
+    const cognitoService = new CognitoServiceImpl(
       this.dataDirectory,
       clients,
-      this.clock,
       userPoolDefaultConfig,
-      this.userPoolServiceFactory
+      this.userPoolServiceFactory,
     );
+
+    await cognitoService.init(ctx);
+
+    return cognitoService;
   }
 }
